backend/pyproject.toml
[project]
name = "incident-agent-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.12",
    "uvicorn==0.34.2",
    "sqlalchemy==2.0.40",
    "asyncpg==0.30.0",
    "pydantic[email]==2.11.3",
    "httpx>=0.25.0",
    "greenlet>=3.0.0",
    "python-multipart>=0.0.24",
    "httpx>=0.28.0",
    # LLM providers — install all so anyone can switch via LLM_PROVIDER without reinstalling
    "langchain-core>=0.3",
    "langchain-openai>=0.3",      # covers openai + openrouter (OpenAI-compatible)
    "langchain-anthropic>=0.3",
    "langchain-google-genai>=2.0",
    "langgraph>=0.3",             # agent framework used by the triage agent
    # Observability
    "arize-phoenix-otel>=0.4",
    "opentelemetry-exporter-otlp>=1.29",
    "opentelemetry-instrumentation-fastapi>=0.50",
    "opentelemetry-instrumentation-httpx>=0.50",
    "opentelemetry-instrumentation-sqlalchemy>=0.50",
    "openinference-instrumentation-langchain>=0.1",
]

[dependency-groups]
dev = [
    "ruff>=0.11.0",
    "mypy>=1.15.0",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["B008"]
fixable = ["ALL"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
